Add missing degree to radian conversion
When running the code I saw the moon-earth-sun positions where not on a single line as expected for a full moon day (2022-08-12). After checking the calculations with the example in https://www.stjarnhimlen.se/comp/tutorial.html#5 I found the missing conversion from degrees to radians in celestial_bodies() in solar.cpp.
With thr converion added I get the correct positions for a full moon and the equinoxes.
The sister project for the pico has the conversion included.